+++ BabyBack/src/main/java/com/example/demo/reposity/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
 
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
 
     User findByTel(String tel);
